+++ com.unity.mobile.android-logcat/Editor/UI/AndroidLogcatListViewItem.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e3ce8489378c95c42984e319a8b859ad
+guid: 8472d03b54ff98d4381b12dc87f94bdf
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
