+++ com.unity.mobile.android-logcat/Editor/UI/AndroidLogcatListViewState.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e3e55ab3ff10b724a809adeb77a2af8d
+guid: f78887ddb339d9048838dc1dcea72c83
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
